Fixed Pin Code Feature
Now cross-platform, and stable ( after fixing I tested it a couple dozen times ).
Special Thanks to Yommy for providing me with his packet list, without it I would have never figured what was wrong with it.
Special Thanks to Jeroen who reported a inconsistency on the feature when used on windows.

+++ src/login/login.c
@@ -921,7 +921,7 @@ int parse_fromchar(int fd)
 				
 				if( accounts->load_num(accounts, &acc, RFIFOL(fd,2) ) ) {
 					strncpy( acc.pincode, (char*)RFIFOP(fd,6), 5 );
-					acc.pincode_change = time( NULL );
+					acc.pincode_change = ((unsigned int)time( NULL ));
 					accounts->save(accounts, &acc);
 				}
 				RFIFOSKIP(fd,11);
